+++ src/Admin/AcademicLevelAdmin.php
@@ -17,7 +17,6 @@ final class AcademicLevelAdmin extends AbstractAdmin
         $filter
             ->add('nameFr')
             ->add('nameEn')
-            ->add('academicYear')
         ;
     }
 
@@ -26,7 +25,6 @@ final class AcademicLevelAdmin extends AbstractAdmin
         $list
             ->add('nameFr')
             ->add('nameEn')
-            ->add('academicYear')
             ->add(ListMapper::NAME_ACTIONS, null, [
                 'actions' => [
                     'show' => [],
@@ -41,7 +39,6 @@ final class AcademicLevelAdmin extends AbstractAdmin
         $form
             ->add('nameFr')
             ->add('nameEn')
-            ->add('academicYear')
         ;
     }
 
@@ -50,7 +47,6 @@ final class AcademicLevelAdmin extends AbstractAdmin
         $show
             ->add('nameFr')
             ->add('nameEn')
-            ->add('academicYear')
         ;
     }
 }