+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 🚸 Sort tutoring sessions by start date !29
 - ♻️ Refactor subscribe js fetch flow with await !31
 - 💬 Changed multiple wording and added some notes !32
+- ♻️ Refactor academic year in tutoring instead of academic level !34
 - 🐛 Remove trailing slash on CAS logout redirect url !35
 - ✏️ Changed multiple wording + ui session flex + admin tutoring show !36